% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{afdf}
\alias{afdf}
\title{Shortcut to converting tibbles to data frames}
\usage{
afdf(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Shortcut to converting tibbles to data frames
}
\details{
I'm sick of having to convert tibbles that hide columns back to
  something that doesn't cause unnecessary errors and bugs.
}
