% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_helpers.R
\name{cleanup}
\alias{cleanup}
\title{Attempt to remove temporary files}
\usage{
cleanup(outfile, ask = FALSE)
}
\arguments{
\item{outfile}{String to outfile, including the full path.}

\item{ask}{Logical defaulting to \code{ask = FALSE}. If \code{ask = TRUE},
the function promtps the user with the filenames it has found.}
}
\description{
Attempt to remove temporary files
}
