% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_helpers.R
\name{colorows}
\alias{colorows}
\title{Color the rows in an HTML table}
\usage{
colorows(
  x,
  bgcolor = "yellow",
  color = "blue",
  rows = NULL,
  format = theFormat
)
}
\arguments{
\item{x}{A data frame.}

\item{bgcolor}{Background color, defaulting to "yellow".}

\item{color}{Font color, defaulting to "blue".}

\item{rows}{Row indices to apply the colors to. Defaults to \code{ros = NULL}
and nothing is done.}

\item{format}{String, either "html", "pdf" or "word". If not "html", nothing is done.}
}
\description{
Color the rows in an HTML table
}
\note{
The function colors the cells, not the rows (<tr> tags). It's not ideal,
  but simple.
}
