% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_helpers.R
\name{output_table}
\alias{output_table}
\title{Create a table, depending on the type of output format}
\usage{
output_table(
  x,
  format = theFormat,
  digits = 3,
  row.names = TRUE,
  escape = FALSE,
  align = c("l", rep("r", ncol(x))),
  ...,
  font_size = NULL,
  full_width = NULL
)
}
\arguments{
\item{x}{A data frame. If it is a matrix, it gets turned into a data frame.}

\item{format}{Either "pdf", "html" or "word".}

\item{digits, row.names, escape, align, font_size, full_width}{Passed through (or not)
to kable or flextable.}
}
\description{
Create a table, depending on the type of output format
}
\details{
It kind of works for html and pdf, not really tested for Word. It
  is very limited in terms of the options available.
}
