% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_helpers.R
\name{print.output_table}
\alias{print.output_table}
\title{Print table to HTML, PDF or Word from within a for loop}
\usage{
\method{print}{output_table}(x, format = theFormat)
}
\arguments{
\item{x}{The output of \code{output_table}.}

\item{format}{String: either "html", "pdf" or "word".}
}
\description{
Print table to HTML, PDF or Word from within a for loop
}
\details{
If you use for loops to produce output in an Rmd file, you need to
  explicitly print the outputs, and how that is done can depend on the
  format.
}
