% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_helpers.R
\name{printcv}
\alias{printcv}
\title{Print "DRAFT" and, perhaps, the Git commit version so that the font is white}
\usage{
printcv(format = theFormat)
}
\arguments{
\item{format}{A string, either "html", "pdf", or "word".}
}
\description{
Print "DRAFT" and, perhaps, the Git commit version so that the font is white
}
\details{
If the output format is 'word', anyone can edit it and there's not
  much point putting the commit number in. Also, I can't figure out how to
  make the font white.
}
