% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_html.R
\name{render_html}
\alias{render_html}
\title{This is a fairly pointless function, included only to match the syntax of other render_ functions}
\usage{
render_html(infile, herePath = "Rmd/", output_file = NULL)
}
\arguments{
\item{infile}{String giving the name of the file in the Rmd/ directory,
not including the extension.}

\item{herePath}{String telling here::here where to look. Defaults to
\code{herePath = "Rmd/"}.}

\item{output_file}{Passed to \code{rmarkdown::render}.}
}
\description{
This is a fairly pointless function, included only to match the syntax of other render_ functions
}
