% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSummary.R
\name{fullSummary}
\alias{fullSummary}
\title{Compute a complete set of summary statistics for continuous variables}
\usage{
fullSummary(
  data,
  value = value,
  domain = domain,
  test = test,
  arm = arm,
  visit = visit
)
}
\arguments{
\item{data}{A data.frame containing numeric values and columns for visit and arm, and
optionally containing columns for domain and test.}

\item{value, domain, test, arm, visit}{Unquoted names of columns in \code{data}
to be used in summarizing. They all default to their own names.}
}
\value{
A data frame.
}
\description{
Compute a complete set of summary statistics for continuous variables
}
\details{
Note
  that if 'domain' and 'test' do not exist, they will be set to empty strings.
  If you want a table of all baseline values, not split by arm, you need
  to filter down to the correct visit and replace arm with a single value
  before calling this function. Also NOTE THAT you need to ensure BEFOREHAND that
  missing values are properly included (possibly via \code{dplyr::complete}).
}
\note{
In a sane world, you'd add another line to the pipeline and remove Mean
  and SD. Also, I ought to write kable and formattable methods. Earlier
  versions (prior to 2019-05-22) used strings as all arguments but the first.
  However, dplyr is getting noiser about that, so I switched.
}
